* change ISurface and others types for X,Y,Width,Height from short\ushort to int\uint
git-svn-id: svn://svn.code.sf.net/p/essence-udk/code/trunk@126 950191d2-321e-465b-a8de-7dfce684c78e

+++ EssenceUDK.Platform/DataTypes/Factories/BaseSysFactory.cs
@@ -43,17 +43,17 @@ namespace EssenceUDK.Platform.Factories
             return new BitmapSurface(stream);
         }
 
-        public ISurface CreateSurface(ushort width, ushort height, PixelFormat pixelFormat = PixelFormat.Bpp16A1R5G5B5)
+        public ISurface CreateSurface(uint width, uint height, PixelFormat pixelFormat = PixelFormat.Bpp16A1R5G5B5)
         {
             return new BitmapSurface(width, height, pixelFormat);
         }
 
-        public ISurface CreateSurface(ushort width, ushort height, byte[] bytes, PixelFormat pixelFormat = PixelFormat.Bpp16A1R5G5B5)
+        public ISurface CreateSurface(uint width, uint height, byte[] bytes, PixelFormat pixelFormat = PixelFormat.Bpp16A1R5G5B5)
         {
             return new BitmapSurface(width, height, bytes, pixelFormat);
         }
 
-        public ISurface CreateSurface(ushort width, ushort height, byte* bytes, PixelFormat pixelFormat = PixelFormat.Bpp16A1R5G5B5)
+        public ISurface CreateSurface(uint width, uint height, byte* bytes, PixelFormat pixelFormat = PixelFormat.Bpp16A1R5G5B5)
         {
             return new BitmapSurface(width, height, bytes, pixelFormat);
         }
